cover_letter_generator/config/urls.py
from __future__ import annotations

from django.conf import settings
from django.conf.urls.static import static
from django.contrib import admin
from django.urls import include, path
from django.views.generic import TemplateView

urlpatterns = [
    path("admin/", admin.site.urls),
    path("accounts/", include("allauth.urls")),
    path("", TemplateView.as_view(template_name="home.html"), name="home"),
    path("resumes/", include("cover_letter_generator.apps.resumes.urls")),
    path("jobs/", include("cover_letter_generator.apps.cover_letters.urls")),
    path("dashboard/", include("cover_letter_generator.apps.dashboard.urls")),
]

# Media Assets
urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)

# Development Tools
if settings.DEBUG:
    import debug_toolbar

    urlpatterns += [
        path("__debug__/", include(debug_toolbar.urls)),
        path("__reload__/", include("django_browser_reload.urls")),
    ]
