pyproject.toml
[project]
name = "cover-letter-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backoff>=2.2.1",
    "django-allauth[socialaccount]>=65.3.1",
    "django-extensions>=3.2.3",
    "django-htmx>=1.21.0",
    "django-tailwind-cli>=2.21.1",
    "django>=5.1.5",
    "openai>=1.61.1",
    "pgvector>=0.3.6",
    "psycopg2>=2.9.10",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = ["django-browser-reload>=1.17.0", "django-debug-toolbar>=5.0.1"]


# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
show-fixes = true
target-version = "py312"

line-length = 119
indent-width = 4

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "migrations",
]

[tool.ruff.lint]
extend-select = [
    "B",
    "C4",
    "E",
    "ERA",
    "FLY",
    "G",
    "I",
    "ICN",
    "INP",
    "LOG",
    "N",
    "NPY",
    "PERF",
    "PGH",
    "PIE",
    "PL",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLF",
    "SLOT",
    "T10",
    "T20",
    "TCH",
    "TID",
    "UP",
    "YTT",
]
fixable = ["ALL"]
unfixable = []
ignore = ["PLR0913"]

[tool.ruff.format]
skip-magic-trailing-comma = false
line-ending = "auto"
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports_per_module = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
strict = false
implicit_reexport = true
exclude = ["venv", "migrations"]
